perpusOnline

+++ buku/bukuHapus.php
@@ -1,15 +1,24 @@
 <?php
-include '../koneksi.php';
-$id = $_GET["bukuID"];
+include '../config.php';
 
-    //jalankan query DELETE untuk menghapus data
+// Check if the 'bukuID' key exists in the $_GET array
+if(isset($_GET["bukuID"])) {
+    $id = $_GET["bukuID"];
+
+    // jalankan query DELETE untuk menghapus data
     $query = "DELETE FROM buku WHERE bukuID='$id' ";
     $hasil_query = mysqli_query($koneksi, $query);
 
-    //periksa query, apakah ada kesalahan
+    // periksa query, apakah ada kesalahan
     if(!$hasil_query) {
-      die ("Gagal menghapus data: ".mysqli_errno($koneksi).
-       " - ".mysqli_error($koneksi));
+        die ("Gagal menghapus data: ".mysqli_errno($koneksi).
+            " - ".mysqli_error($koneksi));
     } else {
-      echo "<script>alert('Data berhasil dihapus.');window.location='data_buku.php';</script>";
+        // Handle success case, if needed
+        header("location: data_buku.php");
     }
+} else {
+    // Handle the case where 'bukuID' key is not set
+    echo "Parameter 'bukuID' is missing.";
+}
+?>
